添加Redux thunk。 新增Shift多选
