+++ package.json
@@ -55,8 +55,8 @@
     "react": "^16.12.0",
     "react-addons-update": "^15.6.2",
     "react-app-polyfill": "^1.0.4",
-    "react-color": "^2.18.0",
     "react-async-script": "^1.1.1",
+    "react-color": "^2.18.0",
     "react-content-loader": "^5.0.2",
     "react-dev-utils": "^9.1.0",
     "react-dnd": "^9.5.1",
@@ -71,6 +71,7 @@
     "react-router-dom": "^5.1.2",
     "recharts": "^2.0.0-beta.1",
     "redux": "^4.0.4",
+    "redux-thunk": "^2.3.0",
     "resolve": "1.12.0",
     "resolve-url-loader": "3.1.0",
     "sass-loader": "7.2.0",
